move page tests to seperate folder
