+++ __tests__/logout.test.js
@@ -1,22 +1,22 @@
 import { render, screen } from "@testing-library/react";
 
-import PasswordPage from ".";
+import LogOutPage from "../pages/settings/logout";
 
-jest.mock("../../../context/AuthContext", () => ({
+jest.mock("../context/AuthContext", () => ({
   useAuth: () => ({
     authedUser: {},
   }),
 }));
 
-jest.mock("../../../context/UserContext", () => ({
+jest.mock("../context/UserContext", () => ({
   useUser: () => ({
     user: {},
   }),
 }));
 
 describe("PasswordPage page", () => {
-  beforeEach(() => render(<PasswordPage />));
   it("should render two images", () => {
+    render(<LogOutPage />);
     const images = screen.getAllByRole("images");
     expect(images.length).toBe(2);
   });
