+++ __tests__/sign-in.test.js
@@ -1,14 +1,14 @@
 import { render, screen } from "@testing-library/react";
 
-import SignIn from ".";
+import SignIn from "../pages/sign-in";
 
-jest.mock("../../context/AuthContext", () => ({
+jest.mock("../context/AuthContext", () => ({
   useAuth: () => ({
     authedUser: null,
   }),
 }));
 
-jest.mock("../../context/UserContext", () => ({
+jest.mock("../context/UserContext", () => ({
   useUser: () => ({
     user: null,
   }),
